refactor(ai-gateway): isolate OpenRouter display pricing
