+++ apps/web/src/lib/ai-gateway/providers/openrouter/index.ts
@@ -5,7 +5,6 @@ import {
 } from '@/lib/ai-gateway/models';
 import { isFreeModel } from '@/lib/ai-gateway/is-free-model';
 import PROVIDERS from '@/lib/ai-gateway/providers/provider-definitions';
-import type { StoredModel } from '@kilocode/db';
 import type { OpenRouterModel } from '@/lib/organizations/organization-types';
 import {
   OpenRouterModelsResponseSchema,
@@ -29,9 +28,15 @@ import { isFreeNemotronModel, NVIDIA_TRIAL_TOS } from '@/lib/ai-gateway/provider
 import { applyCustomPricingToModel } from '@/lib/ai-gateway/custom-pricing';
 import { addMonths } from 'date-fns';
 import { isOpenRouterGpt56PromoModel } from '@/lib/ai-gateway/providers/openai';
+import { getModelDisplayPricing } from '@/lib/ai-gateway/providers/openrouter/display-pricing';
 
 // Re-export from shared module for backwards compatibility
 export { normalizeModelId } from '@/lib/ai-gateway/model-utils';
+export {
+  getModelDisplayPricing,
+  undoPricingDiscount,
+} from '@/lib/ai-gateway/providers/openrouter/display-pricing';
+export type { EndpointPricing } from '@/lib/ai-gateway/providers/openrouter/display-pricing';
 
 export function buildAutoModelCatalogEntry(m: AutoModel): OpenRouterModel {
   const input_modalities = ['text'];
@@ -105,31 +110,6 @@ export function formatName(model: OpenRouterModel, preferredIndex: number) {
     }
   }
   return name;
-}
-
-type EndpointPricing = NonNullable<StoredModel['endpoints'][number]['pricing']>;
-
-export function undoPricingDiscount(pricing: EndpointPricing): EndpointPricing {
-  const { discount, ...prices } = pricing;
-  if (discount === undefined || discount <= 0) return pricing;
-  const factor = 1 - discount;
-  if (factor <= 0) return prices;
-  const result = { ...prices };
-  for (const key of Object.keys(prices) as (keyof typeof prices)[]) {
-    const value = prices[key];
-    if (value !== undefined) {
-      result[key] = (Number.parseFloat(value) / factor).toFixed(12);
-    }
-  }
-  return result;
-}
-
-export function getModelDisplayPricing(
-  modelId: string,
-  pricing: EndpointPricing | undefined
-): EndpointPricing | undefined {
-  if (!pricing) return undefined;
-  return isOpenRouterGpt56PromoModel(modelId) ? pricing : undoPricingDiscount(pricing);
 }
 
 export function shouldSuppressOpenRouterModel(model: KiloExclusiveModel): boolean {
